+++ tests/GoogleCloudPubSubAdapterTest.php
@@ -18,6 +18,26 @@ class GoogleCloudPubSubAdapterTest extends TestCase
         $this->assertSame($client, $adapter->getClient());
     }
 
+    public function testGetSetAutoCreateTopics()
+    {
+        $client = Mockery::mock(PubSubClient::class);
+        $adapter = new GoogleCloudPubSubAdapter($client);
+        $this->assertTrue($adapter->areTopicsAutoCreated());
+
+        $adapter->setAutoCreateTopics(false);
+        $this->assertFalse($adapter->areTopicsAutoCreated());
+    }
+
+    public function testGetSetAutoCreateSubscriptions()
+    {
+        $client = Mockery::mock(PubSubClient::class);
+        $adapter = new GoogleCloudPubSubAdapter($client);
+        $this->assertTrue($adapter->areSubscriptionsAutoCreated());
+
+        $adapter->setAutoCreateSubscriptions(false);
+        $this->assertFalse($adapter->areSubscriptionsAutoCreated());
+    }
+
     public function testPublishWhenTopicMustBeCreated()
     {
         $topic = Mockery::mock(Topic::class);
@@ -49,7 +69,7 @@ class GoogleCloudPubSubAdapterTest extends TestCase
         $topic->shouldReceive('exists')
             ->once()
             ->andReturn(true);
-        $topic->shouldNotReceive('create');
+        $topic->shouldNotHaveReceived('create');
         $topic->shouldReceive('publish')
             ->with([
                 'data' => 'a:1:{s:5:"hello";s:5:"world";}',
@@ -63,6 +83,28 @@ class GoogleCloudPubSubAdapterTest extends TestCase
             ->andReturn($topic);
 
         $adapter = new GoogleCloudPubSubAdapter($client);
+
+        $adapter->publish('channel_name', ['hello' => 'world']);
+    }
+
+    public function testPublishWhenAutoTopicCreationIsDisabled()
+    {
+        $topic = Mockery::mock(Topic::class);
+        $topic->shouldNotHaveReceived('exists');
+        $topic->shouldNotHaveReceived('create');
+        $topic->shouldReceive('publish')
+            ->with([
+                'data' => 'a:1:{s:5:"hello";s:5:"world";}',
+            ])
+            ->once();
+
+        $client = Mockery::mock(PubSubClient::class);
+        $client->shouldReceive('topic')
+            ->with('channel_name')
+            ->once()
+            ->andReturn($topic);
+
+        $adapter = new GoogleCloudPubSubAdapter($client, false);
 
         $adapter->publish('channel_name', ['hello' => 'world']);
     }
@@ -118,7 +160,7 @@ class GoogleCloudPubSubAdapterTest extends TestCase
         $topic->shouldReceive('exists')
             ->once()
             ->andReturn(true);
-        $topic->shouldNotReceive('create');
+        $topic->shouldNotHaveReceived('create');
         $topic->shouldReceive('subscription')
             ->with('channel_name')
             ->once()
@@ -172,7 +214,7 @@ class GoogleCloudPubSubAdapterTest extends TestCase
         $subscription->shouldReceive('exists')
             ->once()
             ->andReturn(true);
-        $subscription->shouldNotReceive('create');
+        $subscription->shouldNotHaveReceived('create');
         $subscription->shouldReceive('pull')
             ->once()
             ->andReturn($messageBatch1);
@@ -193,7 +235,7 @@ class GoogleCloudPubSubAdapterTest extends TestCase
         $topic->shouldReceive('exists')
             ->once()
             ->andReturn(true);
-        $topic->shouldNotReceive('create');
+        $topic->shouldNotHaveReceived('create');
         $topic->shouldReceive('subscription')
             ->with('channel_name')
             ->once()
@@ -206,6 +248,79 @@ class GoogleCloudPubSubAdapterTest extends TestCase
             ->andReturn($topic);
 
         $adapter = new GoogleCloudPubSubAdapter($client);
+
+        $handler1 = Mockery::mock(\stdClass::class);
+        $handler1->shouldReceive('handle')
+            ->with(['hello' => 'world'])
+            ->once();
+        $handler1->shouldReceive('handle')
+            ->with('this is a string')
+            ->once();
+
+        $adapter->subscribe('channel_name', [$handler1, 'handle']);
+    }
+
+    public function testSubscribeWhenAutoTopicCreationIsDisabled()
+    {
+        $messageBatch1 = [
+            [
+                'ackId' => 1,
+                'message' => [
+                    'data' => base64_encode('a:1:{s:5:"hello";s:5:"world";}')
+                ],
+            ],
+            [
+                'ackId' => 2,
+                'message' => [
+                    'data' => base64_encode('this is a string')
+                ],
+            ],
+        ];
+        $messageBatch2 = [
+            [
+                'ackId' => 3,
+                'message' => [
+                    'data' => base64_encode('unsubscribe')
+                ],
+            ],
+        ];
+
+        $subscription = Mockery::mock(Subscription::class);
+        $subscription->shouldNotHaveReceived('exists');
+        $subscription->shouldNotHaveReceived('create');
+        $subscription->shouldReceive('pull')
+            ->once()
+            ->andReturn($messageBatch1);
+        $subscription->shouldReceive('acknowledge')
+            ->with(1)
+            ->once();
+        $subscription->shouldReceive('acknowledge')
+            ->with(2)
+            ->once();
+        $subscription->shouldReceive('pull')
+            ->once()
+            ->andReturn($messageBatch2);
+        $subscription->shouldReceive('acknowledge')
+            ->with(3)
+            ->once();
+
+        $topic = Mockery::mock(Topic::class);
+        $topic->shouldReceive('exists')
+            ->once()
+            ->andReturn(true);
+        $topic->shouldNotHaveReceived('create');
+        $topic->shouldReceive('subscription')
+            ->with('channel_name')
+            ->once()
+            ->andReturn($subscription);
+
+        $client = Mockery::mock(PubSubClient::class);
+        $client->shouldReceive('topic')
+            ->with('channel_name')
+            ->once()
+            ->andReturn($topic);
+
+        $adapter = new GoogleCloudPubSubAdapter($client, true, false);
 
         $handler1 = Mockery::mock(\stdClass::class);
         $handler1->shouldReceive('handle')