+++ changelog.md
@@ -3,8 +3,7 @@
 ## 1.0.1 - ?
 
 * Ack messages individually after callable returns successfully
-
-* Initial release
+* Add functionality to enable/disable auto topic & subscription creation
 
 ## 1.0.0 - 2016-09-05